pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-integrity-api"
version = "1.0.0"
description = "Media Structural Integrity Engine™ — Django REST API Backend"
requires-python = ">=3.11,<3.13"
dependencies = [
    # Django stack
    "django>=5.1,<5.2",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-cors-headers>=4.3,<5.0",
    "django-filter>=24.0,<25.0",
    "django-storages>=1.14,<2.0",
    "drf-spectacular>=0.27,<1.0",

    # Task queue
    "celery[redis]>=5.4,<6.0",
    "redis>=5.0,<6.0",
    "flower>=2.0,<3.0",

    # Google APIs
    "google-ads>=23.0.0,<24.0",
    "google-auth>=2.0,<3.0",
    "google-auth-oauthlib>=1.0,<2.0",
    "google-analytics-data>=0.18",
    "google-analytics-admin>=0.22",
    "google-cloud-bigquery>=3.20",

    # Data processing
    "pandas>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",

    # Reporting
    "jinja2>=3.0,<4.0",
    "weasyprint>=68.0,<70.0",
    "openpyxl>=3.1,<4.0",

    # Logging
    "structlog>=23.0,<25.0",

    # Security
    "cryptography>=41.0,<44.0",

    # Database
    "psycopg2-binary>=2.9,<3.0",

    # Server
    "gunicorn>=21.0,<23.0",

    # MCP (Model Context Protocol)
    "mcp[cli]>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "ruff>=0.1",
    "factory-boy>=3.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py"]
